LayoutTests/fast/repaint/background-attachment-local-scroll-expected.txt

Test that a scroll of an overflow scrolling element, with `background-attachment:local`, repaints

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS layerTreeAsText.indexOf('repaint rect') > -1 is true
PASS successfullyParsed is true

TEST COMPLETE
text here
